move test inside "test" folder
