+++ .gitignore
@@ -2,7 +2,6 @@
 .nyc_output
 build
 node_modules
-test
 src/**.js
 coverage
 *.log
