--- CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
-
-### 4.0.1 (2023-05-23)
