+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@vulpemventures/secp256k1-zkp",
-  "version": "4.0.1",
+  "version": "4.0.0",
   "description": "Essential methods of the secp256k1-zkp lib exported to TS/JS for handling Elements confidential transactions",
   "main": "build/main/index.js",
   "typings": "build/main/index.d.ts",
@@ -11,12 +11,13 @@
   "scripts": {
     "compile": "bash ./scripts/compile_wasm_docker",
     "build": "run-p build:*",
-    "build:main": "tsc -p tsconfig.json",
-    "build:module": "tsc -p tsconfig.module.json",
+    "build:main": "tsc -p tsconfig.prod.json",
+    "build:module": "tsc -p tsconfig.prod.module.json",
     "fix": "run-s fix:*",
     "fix:prettier": "prettier \"src/**/*.ts\" --write",
     "fix:lint": "eslint src --ext .ts --fix",
-    "test": "run-s build test:*",
+    "test": "run-s test:*",
+    "test:build": "tsc -p tsconfig.json",
     "test:lint": "eslint src --ext .ts",
     "test:prettier": "prettier \"src/**/*.ts\" --list-different",
     "test:unit": "nyc --silent ava",
