+++ src/test/ecc.spec.ts
@@ -1,10 +1,10 @@
 import anyTest, { TestInterface } from 'ava';
 
-import fixtures from '../fixtures/ecc.json';
+import { loadSecp256k1ZKP } from '../lib/cmodule';
+import { ecc } from '../lib/ecc';
+import { ZKP } from '../lib/interface';
 
-import { loadSecp256k1ZKP } from './cmodule';
-import { ecc } from './ecc';
-import { ZKP } from './interface';
+import fixtures from './fixtures/ecc.json';
 
 const fromHex = (hex: string) => Buffer.from(hex, 'hex');
 const toHex = (buf: Uint8Array) => Buffer.from(buf).toString('hex');
