+++ src/test/fixtures/rangeproof.json
