+++ tsconfig.prod.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig",
+  "exclude": [
+    "node_modules/**",
+    "src/test"
+  ]
+}
